Make 'git bz file <commit>' just file that commit
Deviate from git-format-patch in the handling of a single revision:
file a bug with a single patch is definitely more common than filing
a bug with a range of patches and this results in less mystical
invocations. Hopefully git-format-patch will eventually catch up.

Remove the support for --num=<N> and -N since it's minimally useful
now; it's not impossible to want to do 'git bz file old_commit^.. -2'
but I don't want to clutter the interface.