Improve error handling when bug reference is incorrect
If the user specifies the bug url on the command line but the url violates our
assumptions, git-bz doesn't currently handle it well.  For instance, the
following commandline invocation:

git bz attach http://bugzilla.foo.org HEAD

will fail the regex that checks whether it is a bugzilla url (since it is
missing show_bug.cgi), so it will interpret it as a 'alias:1234'-style
reference.  'http' is interpreted as the alias, but when that is not found in
the git config, it will attempt to look up a browser cookie for the host 'http'.
That will also fail, but it will report a misleading error which is something like:

Please log in to 'http' in Firefox

This patch just does a little sanity checking that the bug id is a number and
aborts with a slightly more helpful error message if that is not true.