File bugs via XML-RPC when possible
If the server supports XML-RPC (3.0 or newer), then filing via
xmlrpc.cgi has advantages:

 - We get a string error message instead of an HTML page
 - We can not specify priority/severity and let the server
   default it; this is important since there aren't standard
   priorities.

So, we try to first file via xmlrpc.cgi's Bug.create method and
if we get a 404, fall back to the old method of a form post.

The configuration for a couple of fields is changed to use the human
readable field names from Bug.create rather than the database-field
names of post_bug.cgi.

 default-bug-severity => default-severity
 default-rep-platform => default-platform