Add support for bugzilla installations in non-root paths
Currently, git-bz assumes that all bugzilla installations are at the root of the
host (e.g. http://foo.com/).  It is currently impossible to use git-bz with a
bugzilla installation that is hosted at e.g. foo.com/bugzilla/.  This patch adds
that ability.  A new configuration option ('bz-tracker.alias.path') allows the
path to be specified for a host.