Fix exception when parsing an alias as a url
In BugHandle.parse(), we first try to parse the bug reference as a url and then
fall back to interpreting it as an alias from the config file.  Unfortunately,
something like "gnome:123456" gets parsed as a url (with a scheme of 'gnome' and
a path of '123456', and no hostname).  This resulted in us passing a None
hostname to tracker_get_auth_user(), which resulted in an uncaught exception.
From now on, only proceed with treating the bug reference as a url if the scheme
is parsed as 'http' or 'https'.