docs: Quote example url to workaround asciidoc bug
The example url in the section "Ways to refer to a bug" was causing
asciidoc to emit invalid xml (by closing a docbook ulink tag before the
closing semicolon for a &gt; replacement giving &gt</ulink>;) which
meant that xmlto could not then produce a man page. Quoting the url
seems to workaround this issue.

asciidoc version was 8.4.4

https://bugzilla.gnome.org/show_bug.cgi?id=621580